requirements.txt
mariadb==1.1.13
mysql-connector-python==9.4.0
packaging==25.0
psutil==7.0.0
python-dotenv==1.1.1
